Fix codex default naming and general code tidy up.
